server/requirements.txt
Flask
Flask-Cors
Pillow
Exif
numpy
scipy
opencv-python